Copy of github.com/JeffyCN/mirrors.git, branch "tools"
This is a copy of the code at commit
e3cdef0084c4c12e73ebe672731c967fbf3f75f1 of github.com/JeffyCN/mirrors.git.

The mirrors repository of JeffyCN is a big mess, so it cannot be forked.
It must be copied. The current repo should be replaced by the correct
source mirror in the future.